Create common TagModal component
